requirements.txt
kafka-python-ng==2.2.3
six==1.16.0 